[Converter] Теперь использует 64 бита для анализа

+++ MainWindow/mainwindow.cpp
@@ -443,7 +443,7 @@ void MainWindow::tools_converter() {
 
     if(m_converter == nullptr) {
         m_converter = new Converter();
-        connect(m_hexview, SIGNAL(valueChanged(uint8_t)), m_converter, SLOT(on_innerValue(uint8_t)));
+        connect(m_hexview, SIGNAL(valueChanged(int64_t)), m_converter, SLOT(on_innerValue(int64_t)));
     }
 
     m_converter->show();
